requirements.txt
requests >= 2.19.0
python-dotenv >= 0.9.1
